Add CI pipeline and pre-commit config
Introduces a comprehensive GitHub Actions CI pipeline for linting, type checking, and running various test suites. Adds a pre-commit configuration for local quality gates, updates .gitignore for coverage.json, and makes minor improvements to logging setup and documentation.

+++ data_processing/data_core.py
@@ -1,7 +1,7 @@
-import logging
-from data_processing.mappings import mappings
 import numpy as np
 import pandas as pd
+
+from data_processing.mappings import mappings
 
 
 def clean_date(df: pd.DataFrame) -> pd.DataFrame:
@@ -15,25 +15,25 @@ def clean_date(df: pd.DataFrame) -> pd.DataFrame:
     pandas.DataFrame: The DataFrame with the cleaned 'data' column.
     """
     replacements = {
-        'purchase in terminal - mobile code': 'terminal purchase',
-        'web payment - mobile code': 'web payment',
-        'transfer from account': 'account transfer',
-        'transfer to account': 'account deposit',
-        'recipient account': 'recipient',
-        'phone number': 'phone',
-        'location: address': 'location',
-        'title': 'description',
-        'payer references': 'references',
-        'orlen': 'Orlen gas station',
-        'starbucks': 'Starbucks coffee shop',
-        'mcd': 'McDonalds restaurant',
-        'netflix': 'Netflix subscription',
-        'investment platform deposit': 'investment deposit',
-        'amazon': 'Amazon shopping',
-        'piotrkowska 157a': 'Biedronka - Piotrkowska 157a',
-        'drewnowska 58a': 'Manufaktura - Drewnowska 58a',
-        'pabianicka 245': 'Port Łódź - Pabianicka 245',
-        'maratonska 24': 'Retkinia Mall - Maratońska 24'
+        "purchase in terminal - mobile code": "terminal purchase",
+        "web payment - mobile code": "web payment",
+        "transfer from account": "account transfer",
+        "transfer to account": "account deposit",
+        "recipient account": "recipient",
+        "phone number": "phone",
+        "location: address": "location",
+        "title": "description",
+        "payer references": "references",
+        "orlen": "Orlen gas station",
+        "starbucks": "Starbucks coffee shop",
+        "mcd": "McDonalds restaurant",
+        "netflix": "Netflix subscription",
+        "investment platform deposit": "investment deposit",
+        "amazon": "Amazon shopping",
+        "piotrkowska 157a": "Biedronka - Piotrkowska 157a",
+        "drewnowska 58a": "Manufaktura - Drewnowska 58a",
+        "pabianicka 245": "Port Łódź - Pabianicka 245",
+        "maratonska 24": "Retkinia Mall - Maratońska 24",
     }
 
     for old, new in replacements.items():
@@ -51,23 +51,23 @@ def process_dataframe(df: pd.DataFrame) -> pd.DataFrame:
     df = clean_date(df)
 
     # Map categories
-    df['category'] = df['data'].map(mappings)
+    df["category"] = df["data"].map(mappings)
 
     # Remove income (positive prices)
-    df['price'] = df['price'].astype(float)
-    df.loc[df['price'] > 0, 'price'] = np.nan
+    df["price"] = df["price"].astype(float)
+    df.loc[df["price"] > 0, "price"] = np.nan
 
     # Remove '-' from price column and convert to string
-    df['price'] = df['price'].astype(str).str.replace("-", "")
+    df["price"] = df["price"].astype(str).str.replace("-", "")
 
     # Reorder columns
-    desired_columns = ['month', 'year', 'price', 'category', 'data']
+    desired_columns = ["month", "year", "price", "category", "data"]
     df = df[desired_columns]
 
     # Drop rows with NaN in 'price'
-    df = df[df['price'] != 'nan'].reset_index(drop=True)
+    df = df[df["price"] != "nan"].reset_index(drop=True)
 
     # Remove unwanted markers from category
-    df['category'] = df['category'].astype(str).str.replace('🔖🔖🔖', '')
+    df["category"] = df["category"].astype(str).str.replace("🔖🔖🔖", "")
 
     return df
